改顏色

+++ prjEShopping/Controllers/SellerShipmentController.cs
@@ -47,7 +47,7 @@ namespace prjEShopping.Controllers
             ViewBag.Pmethod = (db.PaymentMethods.Where(x => x.PaymentMethodId == data.PaymentMethodId).FirstOrDefault()).PaymentMethodName;
             ViewBag.Recriver = data.Receiver;
             ViewBag.RAddress = data.ReceiverAddress;
-            ViewBag.Fright = (db.ShippingMethods.Where(x => x.ShippingMethodId == data.ShippingMethodId).FirstOrDefault()).Freight;
+            ViewBag.Fright = (int)db.ShippingMethods.Where(x => x.ShippingMethodId == data.ShippingMethodId).FirstOrDefault().Freight;
 
 			//購買人
 			var userorderid = db.Shipments.Where(x => x.ShipmentNumber == ShipNum).SingleOrDefault();
@@ -58,6 +58,9 @@ namespace prjEShopping.Controllers
             //濃縮版
 			var username = db.Users.Where(x => x.UserId == (db.Orders.Where(y => y.OrderId == userorderid.OrderId).FirstOrDefault().UserId)).SingleOrDefault().UserName;
             ViewBag.username = username;
+
+            //購買內容
+            
 			return View();
         }
     }
